pyproject.toml
[project]
name = "ai-stock-analysis"
version = "0.1.0"
description = "ระบบวิเคราะห์หุ้น S&P 500 อัตโนมัติ — Buy-side quality research reports"
requires-python = ">=3.11"
dependencies = [
    # Async & HTTP
    "httpx>=0.27",
    "asyncio-throttle>=1.0",
    # Data
    "pandas>=2.2",
    "numpy>=1.26",
    "pandas-ta>=0.4.71b0",
    # Financial data
    "yfinance>=1.3.0",
    "fredapi>=0.5",
    # LLM
    "groq>=0.9",
    # MCP (Model Context Protocol) — SEC EDGAR server
    "mcp>=1.0",
    # Database
    "sqlalchemy>=2.0",
    # Config & Logging
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
    # Templating & PDF
    "jinja2>=3.1",
    "playwright>=1.40",
    # Charts
    "matplotlib>=3.9",
    "seaborn>=0.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5",
]

[tool.hatch.build.targets.wheel]
packages = ["src/stock_analysis"]

[project.scripts]
fetch = "stock_analysis.data.fetch:main"
screen = "stock_analysis.screening.screener:main"
valuate = "stock_analysis.calc_engine.valuate:main"
analyze = "stock_analysis.llm.analyze:main"
run = "stock_analysis.run:main"
sec-mcp-server = "stock_analysis.mcp_servers.sec_edgar_server:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
